+++ source/app/service-providers/updates/index.ts
@@ -23,7 +23,6 @@ import path from 'path'
 import crypto from 'crypto'
 import got, { Response } from 'got'
 import semver from 'semver'
-import { getConverter } from '@common/util/md-to-html'
 
 import { ipcMain, app, shell } from 'electron'
 import { trans } from '@common/i18n-main'
@@ -35,6 +34,8 @@ import LogProvider from '../log'
 import CommandProvider from '../commands'
 import { ServerAPIResponse, UpdateState } from '@dts/main/update-provider'
 import ConfigProvider from '@providers/config'
+import { md2html } from '@common/modules/markdown-utils'
+import { CITEPROC_MAIN_DB } from '@dts/common/citeproc'
 
 const CUR_VER = app.getVersion()
 const REPO_URL = 'https://zettlr.com/api/releases/latest'
@@ -205,14 +206,12 @@ export default class UpdateProvider extends ProviderContract {
       return
     }
 
-    const md2html = getConverter()
-
     // First we need to parse the JSON data.
     const parsedResponse: ServerAPIResponse = JSON.parse(response.body)
 
     this._updateState.tagName = parsedResponse.tag_name
     this._updateState.updateAvailable = semver.lt(CUR_VER, parsedResponse.tag_name)
-    this._updateState.changelog = md2html(parsedResponse.body)
+    this._updateState.changelog = md2html(parsedResponse.body, CITEPROC_MAIN_DB)
     this._updateState.prerelease = parsedResponse.prerelease
     this._updateState.releasePage = parsedResponse.html_url